+++ src/commonMain/kotlin/de/westermann/kobserve/basic/DirectReceiverReadOnlyProperty.kt
@@ -1,5 +1,7 @@
-package de.westermann.kobserve
+package de.westermann.kobserve.basic
 
+import de.westermann.kobserve.EventHandler
+import de.westermann.kobserve.ReadOnlyProperty
 import kotlin.reflect.KProperty0
 
 class DirectReceiverReadOnlyProperty<T>(