+++ src/commonMain/kotlin/de/westermann/kobserve/basic/ReceiverProperty.kt
@@ -1,5 +1,8 @@
-package de.westermann.kobserve
+package de.westermann.kobserve.basic
 
+import de.westermann.kobserve.EventHandler
+import de.westermann.kobserve.Property
+import de.westermann.kobserve.ReadOnlyProperty
 import kotlin.reflect.KMutableProperty1