+++ src/commonMain/kotlin/de/westermann/kobserve/list/ObservableList.kt
@@ -1,0 +1,3 @@
+package de.westermann.kobserve.list
+
+interface ObservableList<T> : ObservableReadOnlyList<T>, MutableList<T>