+++ src/commonTest/kotlin/de/westermann/kobserve/ReceiverPropertyTest.kt
@@ -1,5 +1,6 @@
 package de.westermann.kobserve
 
+import de.westermann.kobserve.basic.mapBinding
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -7,7 +8,7 @@ class ReceiverPropertyTest {
 
     @Test
     fun receiverPropertyTest() {
-        val userProperty = property(Person("John", "Doe"))
+        val userProperty = de.westermann.kobserve.basic.property(Person("John", "Doe"))
         var user by userProperty
 
         val firstNameProperty = userProperty.mapBinding(Person::firstName)